requirements.txt
python-fasthtml==0.12.21
pystac>=1.9.0
shapely>=2.0.0
python-dateutil>=2.8.2
jsonschema>=4.22.0
laspy[lazrs]>=2.5.0
